fix clippy warnings
