+++ src/arp.rs
@@ -6,7 +6,8 @@ use nix::unistd::close;
 use tracing::debug;
 
 use crate::ether::{EthType, EthernetFrame};
-use crate::utils::{get_local_ip_addr, iptobyte};
+use crate::net::get_local_ip_addr;
+use crate::utils::iptobyte;
 
 #[derive(Debug)]
 pub struct Arp {
@@ -22,21 +23,21 @@ pub struct Arp {
 }
 
 fn send_arp(ifname: &str, target_ip: &str) {
-    let (mac_addr, ip_addr, ifindex) = get_local_ip_addr(Some(ifname)).unwrap();
+    let ni = get_local_ip_addr(Some(ifname)).unwrap().unwrap();
     debug!(
         "ip_addr={:?}, target_ip={:?}",
-        ip_addr.unwrap().to_be_bytes().to_vec(),
+        ni.ip_addr.to_be_bytes().to_vec(),
         iptobyte(target_ip)
     );
 
     let ethernet = EthernetFrame::new(
         vec![0xff, 0xff, 0xff, 0xff, 0xff, 0xff],
-        mac_addr.unwrap().to_vec(),
+        ni.mac_addr.to_vec(),
         EthType::Arp,
     );
     let arp_req = Arp::new(
-        mac_addr.unwrap().to_vec(),
-        ip_addr.unwrap().to_be_bytes().to_vec(),
+        ni.mac_addr.to_vec(),
+        ni.ip_addr.to_be_bytes().to_vec(),
         iptobyte(target_ip),
     );
 
@@ -45,9 +46,7 @@ fn send_arp(ifname: &str, target_ip: &str) {
     send_arp.append(&mut arp_req.to_byte_array());
     debug!(?send_arp);
 
-    let arpreply = arp_req
-        .send(mac_addr.unwrap(), send_arp, ifindex.unwrap())
-        .unwrap();
+    let arpreply = arp_req.send(ni.mac_addr, send_arp, ni.ifindex).unwrap();
 }
 
 impl Arp {
